Rhombus Generator
C program that asks for a number between 1 and 9 (it has to ask while the number is outside the range) and prints a rhombus on the screen using the number. Every line will go from 1 to the line number and from to the line number minus one to 1.
The introduced number will always be in the middle of the rhombus.